Did some refactoring of file names and tidied some imports and function calls
