+++ transformationScripts/transform_json_finincial_support_self.py
