+++ transformationScripts/transform_json_lists.py
@@ -1,4 +1,3 @@
-from extract_json import extract_json
 import pandas as pd
 from datetime import datetime
