+++ transformationScripts/transform_json_self_development.py
